WetGasPvt: move implementation to impl file
